+++ include/UDRefl/ReflMngr.h
@@ -44,6 +44,12 @@ namespace Ubpa::UDRefl {
 		void Clear() noexcept;
 
 		//
+		// Traits
+		///////////
+
+		bool ContainsVirtualBase(Type type) const;
+
+		//
 		// Factory
 		////////////
 		//
@@ -102,14 +108,14 @@ namespace Ubpa::UDRefl {
 		// Modifier
 		/////////////
 
-		Type RegisterType(Type type, size_t size, size_t alignment);
+		Type RegisterType(Type type, size_t size, size_t alignment, bool is_polymorphic = false);
 		Name AddField(Type type, Name field_name, FieldInfo fieldinfo);
 		Name AddMethod(Type type, Name method_name, MethodInfo methodinfo);
 		Type AddBase(Type derived, Type base, BaseInfo baseinfo);
 		bool AddTypeAttr(Type type, Attr attr);
 		bool AddFieldAttr(Type type, Name field_name, Attr attr);
 		bool AddMethodAttr(Type type, Name method_name, Attr attr);
-
+		
 		Name AddTrivialDefaultConstructor(Type type);
 		Name AddTrivialCopyConstructor   (Type type);
 		Name AddZeroConstructor          (Type type);
@@ -117,7 +123,7 @@ namespace Ubpa::UDRefl {
 		// -- template --
 
 		// call
-		// - RegisterType(type_name<T>(), sizeof(T), alignof(T))
+		// - RegisterType(type_name<T>(), sizeof(T), alignof(T), std::is_polymorphic<T>)
 		// - details::TypeAutoRegister<T>::run
 		// you can custom type register by specialize details::TypeAutoRegister<T>
 		template<typename T>