+++ doc/change_log.md
@@ -6,6 +6,8 @@
 
 - 0.8.3
   - rename meta
+  - refactor `BaseInfo` 
+  - `TypeInfo` += `bool is_polymorphic` 
 - 0.8.2
   - use concept for traits (operations, container API and flag)
   - let some APIs return fundamental type (`void`, `std::size_t`, `bool`)